Move the BC factory back
