+++ pyiron_continuum/fenics/factory.py
@@ -6,7 +6,8 @@
 Factories for Fenics-related object creation.
 """
 
-from pyiron_base import ImportAlarm
+from pyiron_base import ImportAlarm, PyironFactory, HasStorage
+from pyiron_continuum.fenics.wrappers import BCParser, DirichletBC
 
 with ImportAlarm(
     "fenics functionality requires the `fenics`, `mshr` modules (and their dependencies) specified as extra"
@@ -17,7 +18,6 @@ with ImportAlarm(
     from fenics import (
         near
     )
-from pyiron_base import PyironFactory
 
 __author__ = "Liam Huber, Muhammad Hassani, Niklas Siemer"
 __copyright__ = (
@@ -699,3 +699,27 @@ class SolverConfig:
 class NotSetCorrectlyError(Exception):
     "raised when the object is not configured correctly!"
     pass
+
+
+class BoundaryConditions(PyironFactory, HasStorage):
+    def __init__(self):
+        PyironFactory.__init__(self)
+        HasStorage.__init__(self)
+        self.storage.pairs = []
+
+    def add(self, value, condition):
+        BCParser(value)
+        BCParser(condition)
+        self.storage.pairs.append((value, condition))
+
+    def list(self):
+        return self.storage.pairs
+
+    def clear(self):
+        self.storage.pairs = []
+
+    def pop(self, i):
+        return self.storage.pairs.pop(i)
+
+    def __call__(self, function_space):
+        return [DirichletBC(*args)(function_space) for args in self.storage.pairs]
